Refactoring made
